add greedy option to parse-expression
In some cases we need to parse just a chunk of the input.  It
needs to parse up to and as much as it can and even if there is
stuff still unparsed characters it declares victory anyway because it
was able to parse some piece out of it. In this sense it is "greedy"
as it just grabs whatever it can. This is made to be the default

In other cases, we need to see if a parser can parse only
succeed if it parsed the entire input frontier in its entirety,
and this is when we pass #f for greedy.  A parse-empty is added
to support this to see if anything remains in the input stream

Now with this option we can start off a parse as  non-greedy , like saying
"Ok, this entire line better parse as an expression" (so that for
example if we have a binary expression in the input, then it doesn't
just parse the left side terminal and reeturn) but then allowing
us to be greedy when parsing pieces of an expression for example
when just extracting the left side of a binary expression and you
know that more parsing needs to be done on the right side.