fix parse-binary-exp to also work with a peek
We had broke parse-binary-exp by removing the non-greedy
parsing of the right hand side, which fixed it's usage for parse-if-exp,
but then broke parse-binary-exp parsing of nested expressions.

The right approach to make it parse until it peeks a specific
text in the input is to also give parse-binary-exp the greedy option
and "followed-by" peek function as parameters, just like we gave
parse-expression.

Prior to this fix, had been trying to acheive the "parse a binary exp,
then expect to parse the end text" by constructing a (parse-pipeline ...
parse-binary-exp... followed-by), but that doesn't work because
parse-pipeline expects the first parser to succeed before it can move
onto the next, but in this case the first one would never succeed
because it would be looking for the end of input and not find
it because the next piece of text is in the input... so the second
part of parse-pipeline that is looking for that ending text
would not have got a chance to run.