fix bug where ws not included ON parsed keywords
Needed to fix so that required ws on boundaries
of keywords gets parsed as part of those keywords.
The bug happening here is that ws is getting parsed
separately and entering the parse-tree as a separate entity

Also as part of this commit, cleanup and remove
the concepts of "parse-phrase" and "parse-string".
Those are deprecated now in favor of the new "parse-keyword"